+++ CHANGES.md
@@ -12,6 +12,8 @@ Releases are also tagged in git, if that's helpful.
 
 ## Coming up
 
+- Make `ga` backscraper take kwargs; fix a bug in 2018 #1349
+- Implement extract from text for `ga` #1349
 
 ## Current